+++ waku/persistence/sql_queries.go
@@ -5,7 +5,7 @@ import (
 	"fmt"
 )
 
-// Queries are the postgresql queries for a given table.
+// Queries are the SQL queries for a given table.
 type Queries struct {
 	deleteQuery  string
 	existsQuery  string
@@ -18,12 +18,8 @@ type Queries struct {
 	getSizeQuery string
 }
 
-// NewQueries creates a new Postgresql set of queries for the passed table
-func NewQueries(tbl string, db *sql.DB) (*Queries, error) {
-	err := CreateTable(db, tbl)
-	if err != nil {
-		return nil, err
-	}
+// Note: Do not use this function to create queries for a table, rather use <rdb>.NewQueries to create table as well as queries.
+func CreateQueries(tbl string, db *sql.DB) *Queries {
 	return &Queries{
 		deleteQuery:  fmt.Sprintf("DELETE FROM %s WHERE key = $1", tbl),
 		existsQuery:  fmt.Sprintf("SELECT exists(SELECT 1 FROM %s WHERE key=$1)", tbl),
@@ -34,7 +30,7 @@ func NewQueries(tbl string, db *sql.DB) (*Queries, error) {
 		limitQuery:   ` LIMIT %d`,
 		offsetQuery:  ` OFFSET %d`,
 		getSizeQuery: fmt.Sprintf("SELECT length(data) FROM %s WHERE key = $1", tbl),
-	}, nil
+	}
 }
 
 // Delete returns the query for deleting a row.
@@ -80,14 +76,4 @@ func (q Queries) Offset() string {
 // GetSize returns the query for determining the size of a value.
 func (q Queries) GetSize() string {
 	return q.getSizeQuery
-}
-
-// CreateTable creates the table that will persist the peers
-func CreateTable(db *sql.DB, tableName string) error {
-	sqlStmt := fmt.Sprintf("CREATE TABLE IF NOT EXISTS %s (key TEXT NOT NULL UNIQUE, data BYTEA);", tableName)
-	_, err := db.Exec(sqlStmt)
-	if err != nil {
-		return err
-	}
-	return nil
 }
