+++ waku/persistence/sqlite/sqlite.go
@@ -2,6 +2,7 @@ package sqlite
 
 import (
 	"database/sql"
+	"fmt"
 	"strings"
 
 	"github.com/golang-migrate/migrate/v4/database"
@@ -80,4 +81,23 @@ func Migrate(db *sql.DB) error {
 		return err
 	}
 	return migrate.Migrate(db, migrationDriver, migrations.AssetNames(), migrations.Asset)
+}
+
+// CreateTable creates the table that will persist the peers
+func CreateTable(db *sql.DB, tableName string) error {
+	sqlStmt := fmt.Sprintf("CREATE TABLE IF NOT EXISTS %s (key TEXT NOT NULL UNIQUE, data BLOB);", tableName)
+	_, err := db.Exec(sqlStmt)
+	if err != nil {
+		return err
+	}
+	return nil
+}
+
+// NewQueries creates a table if it doesn't exist and a new SQL set of queries for the passed table
+func NewQueries(tbl string, db *sql.DB) (*persistence.Queries, error) {
+	err := CreateTable(db, tbl)
+	if err != nil {
+		return nil, err
+	}
+	return persistence.CreateQueries(tbl, db), nil
 }
